// back-end/src/api/users/projects/dto/create-project.input.ts
import { InputType, Int, Field } from '@nestjs/graphql';
import {
  Eligibility,
  ProjectRole,
  ProjectStatus,
  RetentionType,
} from 'src/entities/project-details.entity';
import { Group } from 'src/entities/user-details.entity';

@InputType({
  description:
    'Input type for creating a new project under a specific business.',
})
export class CreateProjectInput {
  @Field({ description: 'ID of the business the project belongs to.' })
  company_id: number;

  // @Field()
  // project_id: number; // Optional if auto-generated by the database

  @Field({ description: 'Name of the project.' })
  project_name: string;

  @Field({
    description:
      'Role of the project in the system (e.g., Principal, Head Contractor, Sub Contractor).',
  })
  project_role: ProjectRole;

  @Field({ description: 'Start date of the project.' })
  project_date: Date;

  @Field({ description: 'Brief description of the project scope or purpose.' })
  project_description: string;

  @Field({ description: 'Full site address of the project.' })
  site_address: string;

  @Field({ description: 'Country where the project is located.' })
  country: string;

  @Field({ description: 'State or region of the project location.' })
  region: string;

  @Field({
    description:
      'Google Place ID for the project location, used for geolocation services.',
  })
  place_id: string;

  @Field({ description: 'Latitude coordinate of the project site.' })
  latitude: string;

  @Field({ description: 'Longitude coordinate of the project site.' })
  longitude: string;

  @Field({
    description: 'Head contract sum (total contract value) for the project.',
  })
  head_contract_sum: number;

  @Field({
    description:
      'Retention type applicable to the project (e.g., Fixed, Percentage).',
  })
  retention_type: RetentionType;

  @Field({
    description:
      'Total number of units (e.g., apartments, buildings) associated with the project.',
  })
  number_of_units: number;

  @Field({
    description:
      'Eligibility for PTA (Practical Completion Payment Adjustment).',
  })
  pta_eligibility: Eligibility;

  @Field({ description: 'Eligibility for RTA (Retention or Trust Account).' })
  rta_eligibility: Eligibility;

  @Field({
    description:
      'Current status of the project (e.g., Active, Completed, Archived).',
  })
  project_status: ProjectStatus;

  @Field({
    nullable: true,
    description: 'User ID of the logged-in user creating the project.',
  })
  created_by: number;

  @Field({
    nullable: true,
    description: 'Timestamp indicating when the project record was created.',
  })
  created_on: Date;

  @Field({
    nullable: true,
    description:
      'Group (User, Admin, System) under which the project record was created.',
  })
  created_group: Group;
}